Extract the QuestionPrompter to its own class
